Live Generation of Arrays
removes memory cap for phrases because smaller arrays are generated and
completed one after the other. Next step: put logging back in, ideally
the program would erase the words that are not “words” from the array
then flush them into the file. That would eliminate overhead of
constantly writing to the array.